#99 Separated examples
