+++ .github/workflows/go-test-darwin.yaml
@@ -29,11 +29,11 @@ jobs:
         timeout-minutes: 12
       - name: gRPC service
         run: docker run --detach --env SENZING_TOOLS_ENABLE_ALL --publish 8261:8261 senzing/serve-grpc:latest
-      # - run: go test -v -p 1 ./...
-      - run: go test -v ./.
-      - run: go test -v ./g2config
-      - run: go test -v ./g2configmgr
-      - run: go test -v ./g2diagnostic
+      - run: go test -v -p 1 ./...
+      # - run: go test -v ./.
+      # - run: go test -v ./g2config
+      # - run: go test -v ./g2configmgr
+      # - run: go test -v ./g2diagnostic
       # - run: go test -v ./g2engine
-      - run: go test -v ./g2product
-      - run: go test -v ./helper
+      # - run: go test -v ./g2product
+      # - run: go test -v ./helper
