+++ .github/workflows/go-test-linux.yaml
@@ -28,11 +28,11 @@ jobs:
         uses: actions/setup-go@v4
         with:
           go-version: ${{ matrix.go }}
-      # - run: go test -v -p 1 ./...
-      - run: go test -v ./.
-      - run: go test -v ./g2config
-      - run: go test -v ./g2configmgr
-      - run: go test -v ./g2diagnostic
+      - run: go test -v -p 1 ./...
+      # - run: go test -v ./.
+      # - run: go test -v ./g2config
+      # - run: go test -v ./g2configmgr
+      # - run: go test -v ./g2diagnostic
       # - run: go test -v ./g2engine
-      - run: go test -v ./g2product
-      - run: go test -v ./helper
+      # - run: go test -v ./g2product
+      # - run: go test -v ./helper
