+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -
 
+## [0.4.1] - 2023-10-03
+
+### Changed in 0.4.1
+
+- Updated testing
+
 ## [0.4.0] - 2023-09-26
 
 ### Changed in 0.4.0
